Fixes the Grid snap with a hack for now
Now the ball that is hit the first is passed to snapGrid() and there a
number of locations are calculated and checked against all the balls in
the graph. NB. Ugly for now, but something with a deadline

+++ core/src/com/group66/game/helpers/LevelLoader.java
@@ -27,7 +27,7 @@ public class LevelLoader {
                 String line = s.nextLine();
                 int ypos = Config.BOUNCE_Y_MAX - (2 * linenr + 1) * Config.BALL_RAD;
                 for (int i = 0; i < line.length(); i++) {
-                    int xpos = Config.BOUNCE_X_MIN + (2 * i + 1) * Config.BALL_RAD;
+                    float xpos = Config.BOUNCE_X_MIN + (2 * i + 1) * Config.BALL_RAD;
                     System.out.println("X pos: " + xpos);
 
                     // shift odd rows
